// src/app/generic/model/store.model.ts
export class StoreModel {

  uuid: string;
  city: string;
  postalCode: string;
  street: string;
  street2: string;
  street3: string;
  addressName: string;
  longitude: number;
  latitude: number;
  complexNumber: string;
  showWarningMessage: string;
  todayOpen: string;
  locationType: string;
  collectionPoint: boolean;
  sapStoreID: string;
  todayClose: string;
  distance: number;

}